requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
django==5.2.9
django-tenants==3.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
djoser==2.3.3
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
faker==38.2.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
model-bakery==1.20.5
oauthlib==3.3.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==2.23
pygments==2.19.2
pyjwt==2.10.1
pytest==9.0.1
pytest-django==4.11.1
python-decouple==3.8
python3-openid==3.2.0
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
celery==5.4.0
redis==5.2.1
django-celery-beat==2.9.0
django-extensions==3.2.3
gunicorn==23.0.0
whitenoise==6.9.0
